+++ src/us/kbase/workspace/test/kbase/JSONRPCLayerTest.java
@@ -84,17 +84,18 @@ import com.google.common.collect.ImmutableMap;
  * These tests are specifically for testing the JSON-RPC communications between
  * the client, up to the invocation of the {@link us.kbase.workspace.workspaces.Workspaces}
  * methods. As such they do not test the full functionality of the Workspaces methods;
- * {@link us.kbase.workspace.test.workspaces.TestWorkspaces} handles that. This means
- * that only one backend (the simplest gridFS backend) is tested here, while TestWorkspaces
+ * {@link us.kbase.workspace.test.workspaces.WorkspaceTest} handles that. This means
+ * that only one backend (the simplest gridFS backend) is tested here, while WorkspaceTest
  * tests all backends and {@link us.kbase.workspace.database.WorkspaceDatabase} implementations.
  */
 public class JSONRPCLayerTest extends JSONRPCLayerTester {
 	
 	@Test
 	public void ver() throws Exception {
-		assertThat("got correct version", CLIENT_NO_AUTH.ver(), is("0.8.3-dev1"));
+		assertThat("got correct version", CLIENT_NO_AUTH.ver(), is("0.9.0-dev1"));
 	}
 	
+	@Test
 	public void status() throws Exception {
 		final Map<String, Object> st = CLIENT1.status();
 		System.out.println(st);
@@ -3257,6 +3258,26 @@ public class JSONRPCLayerTest extends JSONRPCLayerTester {
 	}
 	
 	@Test
+	public void adminAuth2Roles() throws Exception {
+		final Map<String, Object> params = new HashMap<>();
+		params.put("command", "createWorkspace");
+		params.put("user", "user3");
+		params.put("params", new CreateWorkspaceParams().withWorkspace("ws"));
+		CLIENT_AA1.administer(new UObject(params)); // has full admin role
+		
+		// has read only role
+		failAdmin(CLIENT_AA3, params, "Full administration rights required for this command");
+		failAdmin(CLIENT_AA2, params, "user2 is not an admin"); // has no role
+		
+		params.put("command", "getWorkspaceInfo");
+		params.put("params", new WorkspaceIdentity().withId(1L));
+		final List<Object> wsinfo = CLIENT_AA3.administer(new UObject(params))
+				.asClassInstance(new TypeReference<List<Object>>() {});
+		assertThat("incorrect ws id", wsinfo.get(0), is(1));
+		assertThat("incorrect ws name", wsinfo.get(1), is("ws"));
+	}
+	
+	@Test
 	public void adminModRequest() throws Exception {
 		Map<String, String> mod2owner = new HashMap<String, String>();
 		checkModRequests(mod2owner);
@@ -3307,14 +3328,14 @@ public class JSONRPCLayerTest extends JSONRPCLayerTester {
 				" \"user\": \"" + USER1 + "\"," +
 				" \"params\": [{\"workspace\": \"" + USER1 + ":admintest\", \"globalread\": \"n\"," +
 				"			   \"description\": \"mydesc\"}]}",
-				"Unable to deserialize CreateWorkspaceParams out of params field.");
+				"Unable to deserialize CreateWorkspaceParams out of params field");
 		
 		failAdmin(CLIENT2,
 				"{\"command\": [\"createWorkspace\"]," +
 				" \"user\": \"" + USER1 + "\"," +
 				" \"params\": {\"workspace\": \"" + USER1 + ":admintest\", \"globalread\": \"n\"," +
 				"			   \"description\": \"mydesc\"}}",
-				"Unable to deserialize a workspace admin command from the input.");
+				"Unable to deserialize a workspace admin command from the input");
 		
 		TypeReference<Tuple9<Long, String, String, String, Long, String, String, String, Map<String, String>>> typeref
 				= new TypeReference<Tuple9<Long, String, String, String, Long, String, String, String, Map<String, String>>>() {};